feat : reward tag 생성 컨트롤러 추가
- reward tag를 생성하는 컨트롤러를 추가했습니다.
